fix: prevent comments in xcconfig (#40)
## Context

When adding url's to the rnuc config setup, they arrive correctly on the Android and JS side, but get cut off on the iOS side. For example, `https://www.github.com` will show up as `https:`. This is because the `//` is interpreted as a comment special character. The most used solution seems to inject `$()` in between the two forward slashes to break them up without actually changing the output string. 

See for example:

- https://stackoverflow.com/questions/21317844/how-do-i-configure-full-urls-in-xcconfig-files
- https://github.com/luggit/react-native-config/pull/374

## Changes

- Adds a new handlebar helper to escape forward slashes.
- Implements the helper in the xcconfig template.
- Updates the test template to reflect the desired outcome.