Add support for react-native-web
This change introduces a new platform in addition to ios and android:
web. An index.web.js file is generated in addition to the existing
output files to support the web environment. The `browser` field of
package.json points to this file, ensuring that web bundlers will pick
it up instead of the index.js file that is intended for a react-native
environment. I've updated the documentation and tests to reflect this
change. A new example app will be added in a subsequent change to avoid
muddying this commit.

This change also adds an example-web package to test
react-native-ultimate-config with a web app. The example-web package was
bootstrapped with `npx create-react-app --use-npm --template typescript`
and modified slightly to depend on react-native-ultimate-config, include
an rnuc script, and import/render a variable from
react-native-ultimate-config.

I also made a slight adjustment to the render-env function so that it
does not fail when run in a web app that does not include an ios folder
(which may be the case for some react-native-web packages).

Fixes #61